Add bla_pogo cvar
Add bla_pogo flag to simulate the behaviour of external scripts like
AutoHotkey to keep jumping repeatedly when the jump button is held down.
Right now the check interval for this is maxed out which means speed
gains occur a lot sooner than with AHK. In the future the jump intervals
should probably be limited or at least customizable through a cvar to
make the system behave more like when AHK was used.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>