Add bla_nodamage console command and...
bla_damagefraction cvar. The command toggles whether the player should take
damage or not. Unlike the `god' command this command does not prevent
any force from being applied to the player as a result of, for instance,
an explosion. The bla_damagefraction cvar can be used to reduce the
amount of damage a player takes compared to normal NPCs.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>