Add bla_infammo console command
Add console command to toggle infinite ammo on/off. This also adds a new
member variable to CBasePlayer which is used as a dedicated flags
variable for blamod now. Previously, we just used CBaseEntity's m_fFlags.
However, since the field is an int which is only four bytes wide when
compiling 32-bit binaries and the SDK alread defines 31 constants we're
bound to run out of available slots at some point. This fix gives us
more than enough room to add blamod-specific flags in the future.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>